requirements.txt
Django==2.2.3
django-datetime-widget==0.9.3
django-phone-field==1.7.1
Pillow==6.1.0
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
